user, transaction and bin api integration started(bug fixed)

+++ app/src/main/java/com/example/smartbin/model/remote/Transaction.kt
@@ -18,7 +18,7 @@ class Transaction {
 
     @SerializedName("coins")
     @Expose
-    var coins: Int? = null
+    var coins: Double? = null
 
     @SerializedName("wasteType")
     @Expose
@@ -26,11 +26,11 @@ class Transaction {
 
     @SerializedName("height")
     @Expose
-    var height: Int? = null
+    var height: Double? = null
 
     @SerializedName("weight")
     @Expose
-    var weight: Int? = null
+    var weight: Double? = null
 
     @SerializedName("status")
     @Expose
